# app/api/openapi.yaml
openapi: 3.0.0
info:
  title: TerraFusion Platform API
  description: API for the TerraFusion geospatial microservices platform.
  version: 1.0.0
  contact:
    name: TerraFusion Support
    email: support@terrafusion.example.com
servers:
  - url: /api
    description: TerraFusion API
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Token:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
      required:
        - access_token
        - token_type
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        before_state:
          type: object
        after_state:
          type: object
        ip_address:
          type: string
        user_agent:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - action
        - resource_type
        - resource_id
        - created_at
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      summary: Login and get access token
      tags:
        - authentication
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Invalid credentials
...
  /audit/logs:
    get:
      summary: Get audit logs with optional filters
      tags:
        - terraaudit
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: user_id
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: resource_type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
  /audit/correction:
    post:
      summary: Submit a data correction
      tags:
        - terraaudit
      parameters:
        - name: feature_id
          in: query
          required: true
          schema:
            type: string
        - name: correction_type
          in: query
          required: true
          schema:
            type: string
        - name: reason
          in: query
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Correction submitted